Add files via upload
This project was created by 20201297 Kim Min-sang and 20201341 Won Chan-hyuk. Please place the image path in C:/Users/.

When implementing the K-means algorithm, we used both the elbow method and the silhouette method to determine the optimal number of clusters and implemented each method.

Another feature of our K-means code is that instead of setting the initial cluster centers randomly, we divided the pixel range of 0-255 into intervals equal to the number of clusters, ensuring that the results remain consistent.

Thank you!